Rename pp_rules to pp_load_rules
