+++ lib/iris/tests/integration/test_pp_constrained_load_cubes.py
@@ -25,7 +25,7 @@ import iris.tests as tests
 
 import iris
 from iris.fileformats import pp
-from iris.fileformats.pp_rules import convert
+from iris.fileformats.pp_load_rules import convert
 from iris.fileformats.rules import load_cubes
