+++ lib/iris/tests/unit/fileformats/pp_rules/test__collapse_degenerate_points_and_bounds.py
@@ -29,7 +29,8 @@ import iris.tests as tests
 
 import numpy as np
 
-from iris.fileformats.pp_rules import _collapse_degenerate_points_and_bounds
+from iris.fileformats.pp_load_rules import \
+    _collapse_degenerate_points_and_bounds
 
 
 class Test(tests.IrisTest):
