+++ lib/iris/tests/unit/fileformats/pp_rules/test__reduced_points_and_bounds.py
@@ -29,7 +29,7 @@ import iris.tests as tests
 
 import numpy as np
 
-from iris.fileformats.pp_rules import _reduce_points_and_bounds
+from iris.fileformats.pp_load_rules import _reduce_points_and_bounds
 
 
 class Test(tests.IrisTest):
